Connect bus-selection sheet with rest of app
