+++ Shared/Views/ContentView.swift
@@ -11,32 +11,6 @@ import Moya
 
 struct ContentView: View {
 	
-	enum StatusText: String {
-		
-		case mapRefresh = "The map automatically refreshes every 5 seconds."
-		case locationData = "You're helping out other users with real-time bus location data."
-		case thanks = "Thanks for helping other users with real-time bus location data!"
-		
-	}
-	
-	let timer = Timer.publish(every: 5, on: .main, in: .common)
-		.autoconnect()
-	
-	var buttonText: String {
-		get {
-			switch self.mapState.travelState {
-			case .onBus:
-				return "Leave Bus"
-			case .notOnBus:
-				return "Board Bus"
-			}
-		}
-	}
-	
-	@State private var statusText = StatusText.mapRefresh
-	
-	@State private var doDisableButton = true
-	
 	@EnvironmentObject private var mapState: MapState
 	
 	@EnvironmentObject private var viewState: ViewState
@@ -81,50 +55,8 @@ struct ContentView: View {
 				}
 				Spacer()
 				#endif // !APPCLIP
-				HStack {
-					Spacer()
-					VStack(alignment: .leading) {
-						Button {
-							switch self.mapState.travelState {
-							case .onBus:
-								self.mapState.busID = nil
-								self.mapState.locationID = nil
-								self.mapState.travelState = .notOnBus
-								self.statusText = .thanks
-								DispatchQueue.main.asyncAfter(deadline: .now() + 5) {
-									self.statusText = .mapRefresh
-								}
-								LocationUtilities.locationManager.stopUpdatingLocation()
-							case .notOnBus:
-								if self.mapState.busID == nil {
-									self.viewState.alertType = .noNearbyBus
-								} else {
-									self.mapState.travelState = .onBus
-									self.statusText = .locationData
-									LocationUtilities.locationManager.startUpdatingLocation()
-								}
-							}
-							self.updateButtonState()
-						} label: {
-							Text(self.buttonText)
-								.bold()
-						}
-							.buttonStyle(BlockButtonStyle())
-							.disabled(self.doDisableButton)
-						HStack {
-							Text(self.statusText.rawValue)
-								.layoutPriority(1)
-							Spacer()
-							self.refreshButton
-								.frame(width: 30)
-						}
-					}
-						.padding()
-						.background(ViewUtilities.standardVisualEffectView)
-						.cornerRadius(20)
-					Spacer()
-				}
-					.padding()
+				PrimaryOverlay()
+					.padding(.bottom)
 				#if APPCLIP
 				Spacer()
 				#endif // APPCLIP
@@ -162,99 +94,70 @@ struct ContentView: View {
 					#else // os(iOS) && !APPCLIP
 					EmptyView()
 					#endif // os(iOS) && !APPCLIP
+				case .busSelection:
+					#if os(iOS)
+					if #available(iOS 15.0, *) {
+						BusSelectionSheet()
+							.interactiveDismissDisabled()
+					} else {
+						BusSelectionSheet()
+					}
+					#else // os(iOS)
+					EmptyView()
+					#endif // os(iOS)
 				}
 			}
 			.alert(item: self.$viewState.alertType) { (alertType) -> Alert in
 				switch alertType {
-				case .noNearbyBus:
+				case .noNearbyStop:
 					let title = Text("No Nearby Stop")
 					let message = Text("You can't board a bus if you're not within ten meters of a stop.")
 					let dismissButton = Alert.Button.default(Text("Continue"))
 					return Alert(title: title, message: message, dismissButton: dismissButton)
 				}
 			}
-			.onReceive(self.timer) { (_) in
-				switch self.mapState.travelState {
-				case .onBus:
-					guard let coordinate = LocationUtilities.locationManager.location?.coordinate else {
-						LoggingUtilities.logger.log(level: .info, "User location unavailable")
-						break
-					}
-					LocationUtilities.sendToServer(coordinate: coordinate)
-				case .notOnBus:
-					guard let location = LocationUtilities.locationManager.location else {
-						break
-					}
-					let closestBus = self.mapState.buses.min { (firstBus, secondBus) -> Bool in
-						let firstBusDistance = firstBus.location.convertForCoreLocation().distance(from: location)
-						let secondBusDistance = secondBus.location.convertForCoreLocation().distance(from: location)
-						return firstBusDistance < secondBusDistance
-					}
-					let closestStopDistance = self.mapState.stops.reduce(into: Double.greatestFiniteMagnitude) { (distance, stop) in
-						let newDistance = stop.location.distance(from: location)
-						if newDistance < distance {
-							distance = newDistance
-						}
-					}
-					if closestStopDistance < 10 {
-						self.mapState.busID = closestBus?.id
-						self.mapState.locationID = UUID()
-					}
-				}
-				self.refreshBuses()
-			}
-	}
-	
-	private var refreshButton: some View {
-		Button(action: self.refreshBuses) {
-			Image(systemName: "arrow.clockwise.circle.fill")
-				.resizable()
-				.aspectRatio(1, contentMode: .fit)
-		}
 	}
 	
 	#if os(macOS)
+	private let timer = Timer.publish(every: 5, on: .main, in: .common)
+		.autoconnect()
+	
 	private var mapView: some View {
 		MapView()
 			.toolbar {
 				ToolbarItem {
-					self.refreshButton
+					Button {
+						NotificationCenter.default.post(name: .refreshBuses, object: nil)
+					} label: {
+						Image(systemName: "arrow.clockwise.circle.fill")
+							.resizable()
+							.aspectRatio(1, contentMode: .fit)
+					}
 				}
 			}
 			.onAppear {
 				NSWindow.allowsAutomaticWindowTabbing = false
 			}
+			.onReceive(NotificationCenter.default.publisher(for: .refreshBuses, object: nil)) { (_) in
+				self.refreshBuses()
+			}
+			.onReceive(self.timer) { (_) in
+				self.refreshBuses()
+			}
+	}
+	
+	private func refreshBuses() {
+		[Bus].download { (buses) in
+			DispatchQueue.main.async {
+				self.mapState.buses = buses
+			}
+		}
 	}
 	#else // os(macOS)
 	private var mapView: some View {
 		MapView()
 	}
 	#endif // os(macOS)
-	
-	func refreshBuses() {
-		[Bus].download { (buses) in
-			DispatchQueue.main.async {
-				self.mapState.buses = buses
-				self.updateButtonState()
-			}
-		}
-//		if let location = locationManager.location {
-//			let locationMapPoint = MKMapPoint(location.coordinate)
-//			let nearestStop = self.mapState.stops.min { (firstStop, secondStop) in
-//				let firstStopDistance = MKMapPoint(firstStop.coordinate).distance(to: locationMapPoint)
-//				let secondStopDistance = MKMapPoint(secondStop.coordinate).distance(to: locationMapPoint)
-//				return firstStopDistance < secondStopDistance
-//			}
-//			let busPoints = self.mapState.buses.map { (bus) -> (bus: Bus, mapPoint: MKMapPoint) in
-//
-//			}
-//			self.statusText = "The next bus is \("?") meters away from the nearest stop."
-//		}
-	}
-	
-	func updateButtonState() {
-		self.doDisableButton = LocationUtilities.locationManager.location == nil || self.mapState.buses.count == 0 && self.mapState.travelState == .notOnBus
-	}
 	
 }
