+++ iOS/ShuttleTrackerApp.swift
@@ -42,7 +42,7 @@ import OnboardingKit
 			.commands {
 				CommandGroup(before: .sidebar) {
 					Button("Refresh") {
-						self.contentView.refreshBuses()
+						NotificationCenter.default.post(name: .refreshBuses, object: nil)
 					}
 						.keyboardShortcut(KeyEquivalent("r"), modifiers: .command)
 				}
