+++ macOS/ShuttleTrackerApp.swift
@@ -39,7 +39,7 @@ import OnboardingKit
 			.commands {
 				CommandGroup(before: .sidebar) {
 					Button("Refresh") {
-						self.contentView.refreshBuses()
+						NotificationCenter.default.post(name: .refreshBuses, object: nil)
 					}
 						.keyboardShortcut(KeyEquivalent("r"), modifiers: .command)
 				}
